+++ integreat_cms/cms/views/pois/poi_form_view.py
@@ -14,7 +14,7 @@ from ...constants import status
 from ...decorators import permission_required
 from ...forms import POIForm, POITranslationForm
 from ...models import POI, POITranslation, Language
-from ...utils.translation_utils import ugettext_many_lazy as __
+from ...utils.translation_utils import translate_link, ugettext_many_lazy as __
 from ..media.media_context_mixin import MediaContextMixin
 from ..mixins import ContentEditLockMixin
 from .poi_context_mixin import POIContextMixin
@@ -177,6 +177,13 @@ class POIFormView(
                 poi_translation_form.instance.poi.translations.filter(
                     language__in=languages
                 ).update(status=status.DRAFT)
+            elif (
+                poi_translation_form.instance.status == status.PUBLIC
+                and poi_translation_form.instance.minor_edit
+            ):
+                poi_translation_form.instance.poi.translations.filter(
+                    language=language
+                ).update(status=status.PUBLIC)
 
             # Show a message that the slug was changed if it was not unique
             if user_slug and user_slug != poi_translation_form.cleaned_data["slug"]:
@@ -184,15 +191,22 @@ class POIFormView(
                     poi__region=region, slug=user_slug, language=language
                 ).first()
                 other_translation_link = other_translation.backend_edit_link
+                message = _(
+                    "The slug was changed from '{user_slug}' to '{slug}', "
+                    "because '{user_slug}' is already used by <a>{translation}</a> or one of its previous versions.",
+                ).format(
+                    user_slug=user_slug,
+                    slug=poi_translation_form.cleaned_data["slug"],
+                    translation=other_translation,
+                )
                 messages.warning(
                     request,
-                    _(
-                        "The slug was changed from '{user_slug}' to '{slug}', because '{user_slug}' is already used by <a href='{link}' class='underline hover:no-underline'>{translation}</a> or one of its previous versions"
-                    ).format(
-                        user_slug=user_slug,
-                        slug=poi_translation_form.cleaned_data["slug"],
-                        link=other_translation_link,
-                        translation=other_translation,
+                    translate_link(
+                        message,
+                        attributes={
+                            "href": other_translation_link,
+                            "class": "underline hover:no-underline",
+                        },
                     ),
                 )