+++ tests/cms/utils/__init__.py
@@ -1,0 +1,3 @@
+"""
+This package contains tests for the :mod:`integreat_cms.cms.utils` package.
+"""