+++ tests/cms/views/view_config.py
@@ -1,10 +1,13 @@
 """
 This modules contains the config for the view tests
 """
+import json
+
 from django.conf import settings
 from django.urls import reverse
 
 from integreat_cms.cms.constants import status
+from integreat_cms.cms.models.pois.poi import get_default_opening_hours
 
 from ...conftest import (
     ALL_ROLES,
@@ -207,6 +210,7 @@ VIEWS = [
                     "longitude": 1,
                     "latitude": 1,
                     "status": status.DRAFT,
+                    "opening_hours": json.dumps(get_default_opening_hours()),
                 },
             ),
             ("pages", STAFF_ROLES + [MANAGEMENT, EDITOR, AUTHOR]),
@@ -244,6 +248,18 @@ VIEWS = [
         ],
         # The kwargs for these views
         {"region_slug": "augsburg", "language_slug": "de"},
+    ),
+    (
+        [
+            ("sbs_edit_imprint", STAFF_ROLES + [MANAGEMENT]),
+            (
+                "sbs_edit_imprint",
+                PRIV_STAFF_ROLES + [MANAGEMENT],
+                {"title": "imprint", "status": status.DRAFT},
+            ),
+        ],
+        # The kwargs for these views
+        {"region_slug": "augsburg", "language_slug": "en"},
     ),
     (
         [
@@ -295,6 +311,7 @@ VIEWS = [
                     "longitude": 1,
                     "latitude": 1,
                     "status": status.DRAFT,
+                    "opening_hours": json.dumps(get_default_opening_hours()),
                 },
             ),
             ("pages", STAFF_ROLES),
@@ -767,6 +784,7 @@ VIEWS = [
                     "longitude": 1,
                     "latitude": 1,
                     "status": status.DRAFT,
+                    "opening_hours": json.dumps(get_default_opening_hours()),
                 },
             ),
             ("archive_poi", PRIV_STAFF_ROLES + REGION_ROLES, {"post_data": True}),